Error reasons for more distribution errors
Add descriptive shutdown reasons when an incoming distribution
connection is rejected because it is not on the "allowed" list or
because it has an incorrect cookie.  These error reasons can be seen
after calling net_kernel:verbose(1).

Distinguishing these failures is valuable, because the log message
printed is otherwise identical.