tools: Optimize adding multiple modules to an Xref server
Extracting data from debug info in BEAM files is parallelized.

The speed-up is not outstanding. If memory is not an issue, adding
`{spawn_opt, [{min_heap_size, <large number>}]}' to the option list of
xref:start/2 can make a (small) difference.

Reading of modules from a directory is parallelized. Attempts to
parallelize all modules when adding a release have been rejected since
the gain showed to be small compared to the increased complexity of
the code.

Using a process for coordinating the reading has not been attempted.