Clean up make_path()
make_path() returns a tuple whose second element is not used by
any of its caller. Eliminate the tuple.

Also avoid calling filename:dirname/1 when we already have the result
stored in a variable, and avoid calling filename:basename/2 on a
complete path when we already have the last part of the pathname
stored in a variable.