Simplify ssl:cipher_suites/1
This PR factors most of the current redundant code into a helper
filtered/1 function, and makes it clearer by inspection that the
`erlang` and `openssl` options actually return the same list of ciphers
with different representations.

As an added benefit, this code no longer relies on the
ssl_cipher:filter_suites/1 function which matches a list of tuples, and
instead only uses the internal binary suite representation.