erl_process: Restore R18 HiPE PCB offsets
It was recently observed that the LLVM backend for HiPE was broken,
generating code that crashes the VM.

It turns out that this was caused by LLVM hard-coding an offset into the
Erlang PCB, which became incorrect as of commit 3ac08f9b, that
introduced a new member into the PCB before the HiPE-specific state,
changing its offsets.

For now, until a proper fix is implemented, we move this new member to
after the HiPE-specific state, allowing the LLVM backend to work once
more.