ssl: Use cypher suite's PRF in prf/5
Use the negotiated cypher suite's PRF algorithm in calls to
ssl:prf/5, rather than a hard-coded one.

For TLS 1.0 the PRF algorithm was hard-coded to MD5/SHA1. This
was correct 100% of the time.

For TLS 1.1 and 1.2 the PRF algorithm was hard-coded to SHA256.
This was correct only some of the time for TLS 1.2 and none of the
time for TLS 1.1. Because the TLS handshake code calls tls_v1:prf/5
through another path, the handshaking process used the negotiated
PRF and did not encounter this bug.

A new test (prf) has been added to ssl_basic_SUITE to guard against future
breakage.

Note that ssl_basic_SUITE:prf_get_ciphers/2 has a cipher filter to work around
a cipher selection bug. See the comments in the code for more information.