Change ct:sleep to timer:sleep in telnet_server
The telnet_server is run on the main test node which will scale and
multiply timers on some test hosts. The other side of the telnet test
(the client) is run on the slave, which does not inherit timer scaling
and multiplication. Therefore, it is better to use timer:sleep in the
server.