hipe: Remove runtime elf_format class switching
The elf_format module was written in such a way that some of the
customisation to ELF-32 vs ELF-64 was made at compile-time and some of
it at run-time. As such it was not actually possible to read 32-bit
files with a module compiled for 64-bit support, or vice versa.

As the run-time selection uses some process dictionary ugliness, it was
removed, shifting all the customisation to be compile-time.