ct_logs: Eliminate dialyzer warnings
common_test uses its own IO server that is group leader for the
test case processes. By default, the IO server escapes characters
with significance to HTML (e.g. '<'). In order to output HTML
tags directly, the IO server must be told not to escape the output.
The way to tell the IO server is to wrap the format string like this:

  ["$ct_html",Format]

That works with common_test's own IO server, but in general not
with other IO servers. Dialyzer will rightly complain that the
call breaks the contract for io:format/3.

To avoid the Dialyzer warning, we must obscure the wrapping of
the format string.

While we are at it, also refactor print_style/4 to make the code
somewhat cleaner and shorter.