test_lib: Correct calculation of number of processes
Correct calculation of the number of parallel processes to use
when cover is used. It was supposed not to exceed 4 (according to
the comment), but it was calculated as the number of schedulers.

On my computer, having 8 schedulers, this change made the compiler
test suite with cover run almost twice as fast.