Tolerate bad directories in the code path
A bad directory in the path would prevent the run-time system
from starting:

$ echo >foobar
$ erl -pa foobar
{"init terminating in do_boot",{load_failed,[supervisor,kernel,gen_server,file_io_server,filename,file,erl_parse,error_logger,code_server,erl_lint,proc_lib,code,application_controller,application_master,gen_event,application,error_handler,lists,heart,gen,file_server,ets,erl_eval]}}

Crash dump is being written to: erl_crash.dump...done
init terminating in do_boot ()

The reason is that when attempting to read each of the BEAM files,
there would be an 'enotdir' error which
erl_prim_load:get_modules/2,3 assumed was a fatal error.

Update erl_prim_load:get_modules/2,3 to ignore any error and try
the next directory in the path.

Reported-by: http://bugs.erlang.org/browse/ERL-142
Reported-by: Michael Truog