beam_type: Correct handling of setelement/3
We must be careful how we treat the type info for the result of:

  setelement(Index, Tuple, NewValue)

If Tuple had type information, the result of setelement/3 (in x(0))
would be assigned the same type information. But that is not safe
for:

  setelement(1, Tuple, NewValue)

since the type for the first element will be changed.

Therefore, we must take care to remove the type information for
the first element of the tuple if might have been modified by
setelement/3.