v3_core: Don't depend on sys_core_fold for cleaning up
a3ec2644f5 attempted to teach v3_core not to generate code with
unbound variables. The approach taken in that commit is to
discard all expressions following a badmatch. That does not
work if the badmatch is nested:

  {[V] = [] = foo,V},
  V

That would be rewritten to:

  {error({badmatch,foo})},
  V

where V is unbound.

If we were to follow the same approach, the tuple construction
code would have to look out for a badmatch. As would list construction,
begin...end, and so on.

Therefore, as it is impractical to discard all expressions that
follow a badmatch, the only other solution is to ensure that the
variables that the pattern binds will somehow be bound. That can
be arranged by rewriting the pattern to a pattern that binds the
same variables. Thus:

  error({badmatch,foo}),
  E = foo,
  case E of
    {[V],[]} ->
      V;
    Other ->
      error({badmatch,Other}
  end