erts: Copy literals in messages on module purge
During check process code, explicitly copy all referenced literals
in a message (in the private queue) to a heap fragment and attach
it to the message reference.

Not all types of message communication does an explicit copy of a
literal and this needs to be taken care of before a module is purged.