Let find-tag visit definitions in right order.
The right order is defined by find-tag-tag-order which is a list of
predicate functions.  Tags matched by the first function shall be
visited first.  After that tags matched by the seconds function shall be
visited and so on.

Previous behavior was to visit tags that match any of the functions in
the order they appear in the TAGS file.

Reported by Dan Gudmundsson.