beam_expect: Correctly handle blocks with multiple allocs
A negative allocation could be calculated if a block had multiple
allocations. Make sure to process the block in the right order
so that the correct allocation is processed. Also add an assertion.

This bug was often not noticed because beam_type usually silently
recalculates the allocation amount in test_heap/2 instructions.