beam_block: Eliminate crash in beam_utils
Somewhat simplified, beam_block would rewrite the target for
the first instruction in this code sequence:

    move x(0) => y(1)
    gc_bif '+' 1 x(0) => y(0)
    move y(1) => x(1)
    move nil => x(0)
    call 2 local_function/2

The resulting code would be:

    move x(0) => x(1)           %% Changed target.
    gc_bif '+' 1 x(0) => y(0)
    move x(1) => y(1)           %% Operands swapped (see 02d6135813).
    move nil => x(0)
    call 2 local_function/2

The resulting code is not safe because the x(1) will be killed
by the gc_bif instruction.

7a47b20c3a cleaned up move optimizations and would reject the
optimization if the target was an X register and an allocating
instruction was found. To avoid this bug, the optimization must be
rejected even if the target is a Y register.