edoc: Provide a slightly better fix for unmatched returns
An attempt to shut off Dialyzer's unmatched return warnings made
a change that simply ignored the return type of a function.
It's arguably better to modify the code of this function to not
return something "by accident."  While at it, rewrote a slighty
confusing use of a list comprehension without a generator to a
case statement.