hipe: Pattern matching compilation of binaries and bistrings
The Core Erlang pattern matching compiler was written long ago, at
a time when binaries and bistrings did not really exist in Erlang.
This patch, taken from the code of CutEr where it's used for more
than a year now, extends the transformation for pattern matching
compilation to also include binaries and bistrings.

Some code that was found erroneous and causes errors when compiling
the transformed code to native code was also taken out while at it.

Thanks to @aggelgian for most of the changes in the code.