make_preload: Save some memory by making preloaded code 'const'
Mark the preloaded code 'const' to allow the compiler to put it into
the 'text' segment instead of into the 'data' segment. Since the
'text' segment is shared among all instances of the Erlang virtual
machine, this change could potentially reduce memory consumption
(slightly).

Before the change:

$ size bin/x86_64-unknown-linux-gnu/beam.smp
   text	   data	    bss	    dec	    hex	filename
2920246	 352273	 158472	3430991	 345a4f	bin/x86_64-unknown-linux-gnu/beam.smp

After the change:

$ size bin/x86_64-unknown-linux-gnu/beam.smp
   text	   data	    bss	    dec	    hex	filename
3081046	 191473	 158472	3430991	 345a4f	bin/x86_64-unknown-linux-gnu/beam.smp

Roughly speaking, this change cuts the size of the data segment in half.