erl_internal: Eliminate duplication of guard tests
Guard tests (e.g. is_list/1) are listed both in guard_bif/2 and
in new_type_test/2. That is a trap for anyone who wants to add
a new guard test. Let guard_bif/2 call new_type_test/2 so that
new guard tests only have to be added in one place.

While we are it, also sort the clauses in guard_bif/2 and
new_type/2.