compiler: Eliminate num_bif_SUITE
num_bif_SUITE.erl was originally copied from the emulator test
suite. It does not test much of the compiler.

Therefore, remove num_bif_SUITE. Add a new test to bif_SUITE
to test trunc/1 and round/1 in contexts that could be tricky
for the compiler to handle correctly. Note that there is no
need to test abs/1 in bif_SUITE, since it is tested in many
other places (e.g. in guard_SUITE).