erts/kernel: Fix code loading deadlock during init:stop
When init:stop is called it walks the application hierarchy
and terminates each process. Some of these processes may do
something while terminating and sometimes that something
needs to load some new code in order to work. When this happens
the code_server could just be in the process of terminating
or the erl_prim_loader could be active. In both these cases
the request to load the new code would cause a deadlock in the
termination of the system.

This commit fixes this by init rejecting attempts to load new code
when init:stop has been called and fixing a termination race in
the code_server.

This however means that the process that tried to do something
when told to terminate (for instance logging that it is terminating)
will crash instead of loading the code.