dialyzer: Suppress warns on generated case stmts
Warnings about clauses that cannot match and are also compiler generated
are suppressed unless none of the clauses return.

This feature is useful for non-Erlang BEAM languages (such as Elixir)
that compile to Erlang and expand certain language constructs into case
statements. In that case, as long as the language construct can succeed,
these warnings are undesired and appear spurious to users that do not
check the Erlang code that their program expands into.

Thanks to José Valim for the test (slightly modified).