[ic] Fix but when UTF-8 character in path to idl spec
The path is only used when writing a comment so the correction
just allowed UTF-8 characters in that io_lib:format call.
IC don't support UTF-8 in the specs or their names
(generated modules contains coding: latin-1) but the path is
only used in a comment so it's not a problem if it can't be
written correctly as long as nothing fails.