kernel: Use ^D as eot for os:cmd on unix platforms
This is needed as doing only an 'exit' will only exit the
program, but any children started in the program that have
stdout/stderr still open will keep the port open until they
terminate. i.e.

    os:cmd("while true; do echo sleep 1; sleep 1; done&").

would block os:cmd forever because the while loop keeps its copies
of stdout/stderr open forever.

It could be argued that this is correct behaviour, and it is
the way it works on windows, but changing this breaks backwards
compatability for os:cmd which is not acceptable.