Improve version selection
Use the list of versions that the server allows and among those choose
the highest version that is not higher than the client's version.

Note that this chosen version might be lower than the client's version,
but is used to improve interoperability.

Patch suggested by Dimitry Borisov refering to RFC 5246 appendix E.1.