erts: Fix internal hashing entropy for maps
We need to use an initial hash seed for each map pair, i.e. the
hash value. The hash value is reset to the seed for each pair instead
of setting the seed to zero.

If we don't, no additional entropy is given to the system and the
hash collision resolution in maps:from_list/1 would fail.