beam_block: Fix potentially unsafe optimization in move_allocates/1
beam_block has an optimization that only is safe when it is applied
immediately after code generation. That is pointed out in a comment:

  NOTE: Moving allocation instructions is only safe because it is done
  immediately after code generation so that we KNOW that if {x,X} is
  initialized, all x registers with lower numbers are also initialized.
  That assumption may not be true after other optimizations, such as
  the beam_utils:live_opt/1 optimization.

The new beam_reorder pass added in OTP 19 runs before beam_block.
Therefore, the optimization is potentially unsafe. The optimization
is also unsafe if compilation is started from assembly code in a
.S file.

Rewrite the optimization to make it safe. See the newly added comment
for details.

ERL-202