Don't crash when obsolete guard overrides local function
The compiler would crash in v3_codegen when trying to compile the
following code:

  is_port(_) -> false.
  foo(P) when port(P) -> ok.

We *could* have the compiler interpret the code as:

  is_port(_) -> false.
  foo(P) when erlang:is_port(P) -> ok.

But that would encourage using the obsolete form of the guard tests.
Note that the following code is illegal:

  is_port(_) -> false.
  foo(P) when is_port(P) -> ok.

It produces the following diagnostic:

  call to local/imported function is_port/1 is illegal in guard

Therefore, we should refuse to compile the code.