erl_internal: Add add_predefined_functions/1
The sys_pre_expand module used to do a lot more (translate records and
funs, for example), but now it does very little. Most of the code is
an identify transformation of the abstract format.

The identity transformation part of the code must be maintained and
kept correct when new forms are added to the abstract format. That
adds to the maintance burden. It also adds (slightly) to compilation
times.  Therefore, we want to eliminate sys_pre_expand, moving all of
its (non-identity) transformations to better places.

As a preliminary first step, move the code that adds the pre-defined
functions (such as module_info/0) to a new function in erl_internal.