Teach erl_expand_records to translate module-less calls
As the next step in eliminating sys_pre_expand, teach
erl_expand_records to handle calls without explicit module name. If
such call refer to a BIF or imported function, add an explicit module
name.

That means that any subsequent pass will know that a call without
a module name is always to a local function defined in the module.